requirements.txt
python-chess==1.999
PyQt5==5.15.11
numpy==1.23.5
torch==2.0.1